requirements.txt
requests==2.26.0
retry==0.9.2
base58==2.1.1
protobuf==3.19.3
grpcio==1.44.0
grpcio-tools==1.44.0
